- Lots of changes:         - base                 - Reorganized file structure                 - Added MeteorOS.Helpers for any global helper functions                   that we want to add         - app-file-browser                 - Added modal for deleting folder that isn't empty         - filesystem                 - Started to add server-side validation of file changes                 - Added owner field

+++ packages/app-file-browser/context-menus/dir.js
@@ -48,7 +48,13 @@ MeteorOS.FS.ContextMenus.Dir = {
             icon: 'glyphicon-trash',
             text: 'Delete Directory',
             action: function(event, selector, context) {
-                context.delete();
+                if (context.files().length > 0) {
+                    MeteorOS.FS.Modals.DeleteFolder(function(result) {
+                        if (result) context.delete();
+                    }).show();
+                } else {
+                    context.delete();
+                }
             }
         },
     ]
